Refactor collision detection
- Drop CollisionDetector class as unnecessary abstraction
- Move collision detection code from area.cpp into area_collision.cpp
- Implement sliding along obstacle normals
- Temporarily disable collision between creatures

+++ src/engine/scene/node/scenenode.h
@@ -32,7 +32,7 @@ namespace reone {
 
 namespace scene {
 
-constexpr int kDefaultDrawDistance = 1024.0f;
+constexpr float kDefaultDrawDistance = 1024.0f;
 
 class SceneGraph;
